dvdnav: fully fix the still image assertion
This completes 24ae7ed55e206a37571d0d39a3d21353936f9a64. Even if the
timer is removed from the schedule, there is no warranty that it's not
pending execution, or waiting on our mutex, at the very moment.
So there is a small window of opportunity for the assertion to fail.

This removes the assertion in favor of a safer if(likely()).
(cherry picked from commit b28c362339ac8ef7f3b9b1715631f304e7563418)