Add a secret --user-agent option to pass the application name
Currently, this is not used anywhere, but better have it for forward
compatibility. This is not as elegant as a dedicated libvlc function
call. But there is no other way to set a parameter before
libvlc_InternalInit() is called by libvlc_new(), short of changing the
libvlc_new() prototype (again). And that would probably be worse.
(cherry picked from commit ae16ebf3524a76da5b2584451bcf0f0f9fc820a4)