Lua Extensions: implement fake "require"
As the "descriptor" function is called with absolutely no libraries
loaded, even the built-in "require" function is not present at scan
time.
This closes #3453.

Note: this is a bit of a hack, but I prefer not to load any libs when
scanning the available extensions

Remove some useless debug, too
(cherry picked from commit 860b78b3d6132b6ab5014a768cbdfbd96c0461fc)

Signed-off-by: Rémi Duraffort <ivoire@videolan.org>