Qt: correctly implement play-and-exit in menu
This calls a var_GetBool each time a menu is redrawn. This may be not
optimal. Maybe a callback on "play-and-exit" would be better?
Do we care about the performance on that part (menu display)?
Close #3842
(cherry picked from commit 5e85bff1849e4391b6598bfbcc5a6c65cd1e151d)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>